Refactor galaxy.tools.exception_handling...
... so datatype module doesn't need to import something from galaxy.tools - which causes a bunch of more things to be imported because galaxy.tools at the top-level has so many imports.
